CAMBIO GRANDE. Distribuido en clases independientes la gestión de ficheros y la de la base de datos. Y creado un gestor para su organización.
